requirements.txt
Django==2.2
djangorestframework==3.9.2
lxml==4.3.3
pytz==2018.9
sqlparse==0.3.0
